Broke TilesInitTask.java out of SubsamplingScaleImageView.java
